requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.30.0
python-dotenv>=1.0.0
tabulate>=0.9.0
click>=8.1.0
pydantic>=2.0.0
